Added the Ghost Model with Animation
The .fbx file contains the model plus animation for the enemy required in the game. The animation is not that much discrete since the ghost model does not have legs and is floating in the air. It also contains the materials for the ghost, viz the normal map and the albedo map.